// src/interface/truck-interface.tsx
export interface Truck {
  registrationNumber: string;
  driverPerformanceGrade: string;
  driverAge: number;
  numberOfViolations: number;
  averageMonthlyDriveTime: string;
  annualAverageDistance: string;
  regularlyServiced: boolean;
  yearOfTruckProduction: number;
  modelOfTruck: string;
  engineType: string;
  cost: number;
}